move fatal_printf logging to glog.c file with other logging that does the same thing.
