+++ src/glog.c
@@ -4,11 +4,19 @@
    See COPYRIGHT.txt for Copyright information
    See LICENSE.txt for license (GPL v2)
  */
-#include <stdio.h>
-#include <time.h>
 #include <stdarg.h>
+#include <stdio.h>
+#include <stdlib.h>
+#include <string.h>
+#include <time.h>
 
 #include "glog.h"
+
+
+#define MAX_FATAL_LOGS          20
+
+int g_fatal_log = 0;
+char *g_fatal_log_strs[MAX_FATAL_LOGS];
 
 int glog(const char *s) {
   time_t timer;
@@ -44,4 +52,41 @@ int glogf(const char *fmt, ...) {
   va_end(ap);
   fputc('\n', stdout);
   return 0;
+}
+
+
+int fatal_printf(const char *fmt, ...)     {
+  static char buffer[4096];
+  va_list ap;
+  int ret;
+
+  va_start(ap, fmt);
+
+  if(g_fatal_log < 0) {
+    g_fatal_log = 0;
+  }
+
+  ret = vsnprintf(buffer, sizeof(buffer), fmt, ap);
+
+  glog(buffer);
+
+  if (g_fatal_log < MAX_FATAL_LOGS) {
+    g_fatal_log_strs[g_fatal_log++] = strdup(buffer);
+  }
+
+
+  va_end(ap);
+  return ret;
+}
+
+
+
+void clear_fatal_logs()      {
+  int i;
+
+  for(i = 0; i < g_fatal_log; i++) {
+    free(g_fatal_log_strs[i]);
+    g_fatal_log_strs[i] = 0;
+  }
+  g_fatal_log = 0;
 }
